remove entities table and change entity guid from integer to 24-char string with type prefix, random chars; switch to amazon SES

+++ engine/widget/news.php
@@ -32,7 +32,7 @@ class Widget_News extends Widget
     
     function render_view($args = null)
     {
-        $end_guid = (int)get_input('end');    
+        $end_guid = get_input('end');    
         return view("widgets/news_view", array('widget' => $this, 'end_guid' => $end_guid));
     }
