+++ mod/contact/schema/mysql.php
@@ -4,7 +4,7 @@ CREATE TABLE `email_templates` (
   `subject` text default null,
   `from` text default null,
   `num_sent` int(11) not null default 0,
-  `time_last_sent` int(11) null,
+  `time_last_sent` bigint(20) unsigned null,
   `filters_json` text default null
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
 
@@ -12,6 +12,6 @@ CREATE TABLE `sms_templates` (
    <?php require 'schema/entity_columns.php'; ?>,
    <?php require 'schema/content_columns.php'; ?>, 
   `num_sent` int(11) not null default 0,
-  `time_last_sent` int(11) null,
+  `time_last_sent` bigint(20) unsigned null,
   `filters_json` text default null
 ) ENGINE=MyISAM DEFAULT CHARSET=utf8;
